+++ Tests/LibGit2Test/LibGit2Test.psm1
@@ -60,6 +60,11 @@ function New-GitTestRepo
 
     $testDrive = (Get-Item -Path 'TestDrive:').FullName
     $repoRoot = Join-Path -Path $testDrive -ChildPath ('LibGit2.{0}' -f ([IO.Path]::GetRandomFileName()))
-    git init $repoRoot | Write-Debug
+    New-GitRepository -Path $repoRoot | Format-List | Out-String | Write-Debug
     return $repoRoot
+}
+
+function Resolve-TestDrivePath
+{
+    return (Get-Item -Path 'TestDrive:').FullName
 }