+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
  * Created `Get-GitRepository` function for getting an object representing a repository.
  * Created `Save-GitChange` function for commiting changes to a repository. Implements the `git commit` command.
  * Created `Find-GitRepository` function for searching a directory and its parents (i.e. up it tree) for a Git repository.
+ * Created `New-GitRepository` function for creating new Git repositories.
 
  
 # 0.0.0 (31 August 2016)